requirements.txt
# API
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# DB
sqlalchemy>=2.0.0

# IA
openai>=1.30.0
anthropic>=0.28.0
google-generativeai>=0.7.0

# Scheduler
apscheduler>=3.10.4
pytz>=2024.1

# Tests
pytest>=8.0.0
httpx>=0.27.0
